requirements.txt
flask
google-generativeai
python-dotenv
numpy
pandas
scikit-learn